pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poly-ai-trading-agent"
version = "2.0.0"
description = "AI trading agents for Polymarket prediction markets"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Polymarket Agents Contributors"}
]
keywords = ["polymarket", "trading", "ai", "prediction-markets", "blockchain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "aiohttp>=3.11.0",
    "chromadb>=0.5.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.1.2",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.2.0",
    "newsapi-python>=0.2.7",
    "openai>=1.54.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "pymongo>=4.6.0",
    "py-clob-client>=0.17.5",
    "py-order-utils>=0.3.2",
    "tavily-python>=0.5.0",
    "typer>=0.12.0",
    "uvicorn[standard]>=0.32.0",
    "web3>=6.20.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=3.8.0",
    "types-requests>=2.32.0",
]

[project.scripts]
poly-agent = "scripts.python.cli:app"

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "ARG001", # unused function argument (common in overrides)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG", "S101"]

[tool.ruff.isort]
known-first-party = ["agents", "scripts"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "langchain.*",
    "langchain_community.*",
    "langchain_core.*",
    "langchain_openai.*",
    "newsapi.*",
    "py_clob_client.*",
    "py_order_utils.*",
    "tavily.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

